Deletado dois dumps que estavam no projeto mas não eram mais usados. Menu está responsivo dependendo de quem está online.

+++ src/main/java/br/com/sqlScholar/controller/StudentController.java
@@ -3,6 +3,8 @@ package br.com.sqlScholar.controller;
 import br.com.sqlScholar.model.Student;
 import br.com.sqlScholar.repository.StudentRepository;
 import br.com.sqlScholar.service.StudentService;
+import jakarta.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.util.DigestUtils;
 import org.springframework.web.bind.annotation.*;
@@ -35,7 +37,11 @@ public class StudentController {
     }
 
     @GetMapping("/index")
-    public ModelAndView index(){
+    public ModelAndView index(HttpSession session){
+        boolean logged = studentService.verifySession(session);
+        if (!logged) {
+            return new ModelAndView("redirect:/");
+        }
         Map<String, Object> template = new HashMap<>();
         int pageNumber = 0;
         int pageSize = 15;
@@ -45,9 +51,20 @@ public class StudentController {
     }
 
     @GetMapping("/perfil/{id}")
-    public ModelAndView perfil(@PathVariable UUID id){
-        Optional<Student> student = this.studentRepository.findById(id);
+    public ModelAndView perfil(@PathVariable UUID id, HttpSession session){
+        boolean logged = studentService.verifySession(session);
+        if (!logged) {
+            return new ModelAndView("redirect:/");
+        }
         Map<String, Object> template = new HashMap<>();
+        if ("admin".equals(session.getAttribute("userType"))) {
+            Optional<Student> student = this.studentRepository.findById(id);
+            template.put("student", student.get());
+        }else{
+            Student student = (Student) session.getAttribute("userLogged");
+            template.put("student", student);
+        }
+
         List<Integer> xValues = new ArrayList<>();//Será substituido por acesso ao banco e dados que virão de lá
         List<Integer> yValues = new ArrayList<>();//Será substituido por acesso ao banco e dados que virão de lá
         int count_solved = (int)(Math.random() * 30);//Será substituido por acesso ao banco e dados que virão de lá (Contagem de questões resolvidas pelo aluno)
@@ -55,7 +72,7 @@ public class StudentController {
             xValues.add((int)(Math.random() * 200));
             yValues.add((int)(Math.random() * 100));
         }
-        template.put("student", student.get());
+       
         template.put("xvalues", xValues.toString());
         template.put("yvalues", yValues.toString());
         template.put("solvedQuestions", count_solved);
@@ -71,7 +88,7 @@ public class StudentController {
         template.put("message", "Aluno criado com sucesso!");
         template.put("arrStudent", this.studentRepository.listAll());
 
-        return new ModelAndView("student/message", template);
+        return new ModelAndView("redirect:/student/perfil/"+student.getId(), template);
     }
 
     @RequestMapping("/editar")
@@ -89,11 +106,15 @@ public class StudentController {
         template.put("message", "Aluno editado com sucesso!");
         template.put("arrStudent", this.studentRepository.listAll());
 
-        return new ModelAndView("student/index", template);
+        return new ModelAndView("redirect:/student/perfil/"+id, template);
     }
 
     @GetMapping("/tela_editar/{id}")
-    public ModelAndView tela_editar(@PathVariable UUID id){
+    public ModelAndView tela_editar(@PathVariable UUID id, HttpSession session){
+        boolean logged = studentService.verifySession(session);
+        if (!logged) {
+            return new ModelAndView("redirect:/");
+        }
         Map<String, Object> template = new HashMap<>();
         Optional<Student> student = this.studentRepository.findById(id);
         template.put("student", student.get());
@@ -101,13 +122,17 @@ public class StudentController {
     }
 
     @GetMapping("/deletar/{id}")
-    public ModelAndView deletar(@PathVariable UUID id){
+    public ModelAndView deletar(@PathVariable UUID id, HttpSession session){
+        boolean logged = studentService.verifySession(session);
+        if (!logged) {
+            return new ModelAndView("redirect:/");
+        }
         this.studentRepository.deleteById(id);
         Map<String, Object> template = new HashMap<>();
         template.put("message", "Aluno deletado com sucesso!");
         template.put("arrStudent", this.studentRepository.listAll());
 
-        return new ModelAndView("student/index", template);
+        return new ModelAndView("redirect:/", template);
     }
 
 }
